MicroInt applied in fiats

+++ app/views/companion/blockchain/IssueAsset.scala
@@ -21,6 +21,6 @@ object IssueAsset {
     )(Data.apply)(Data.unapply)
   )
 
-  case class Data(from: String, to: String, documentHash: String, assetType: String, assetPricePerUnit: Int, quantityUnit: String, assetQuantity: Int, moderated: Boolean, takerAddress: String, gas: Int, mode: String, password: String)
+  case class Data(from: String, to: String, documentHash: String, assetType: String, assetPricePerUnit: Double, quantityUnit: String, assetQuantity: Int, moderated: Boolean, takerAddress: String, gas: Int, mode: String, password: String)
 
 }
